requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Pillow==10.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.1
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1
